requirements.txt
alembic==1.13.3
APScheduler==3.10.4
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
Mako==1.3.6
MarkupSafe==3.0.2
pillow==11.0.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-engineio==4.10.1
python-pptx==1.0.2
python-socketio==5.11.4
pytz==2024.2
requests==2.32.3
simple-websocket==1.1.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.0.4
wsproto==1.2.0
WTForms==3.2.1
XlsxWriter==3.2.0
